Introduce DeleteAction
